+++ packages/orq-rc/src/orq_ai_sdk/models/retrievecontactop.py
@@ -103,7 +103,7 @@ class RetrieveContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-19T05:54:54.980Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-19T11:22:40.604Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")